split tests for existing/default access controller out into separate test class
